break out bottom row and clean up
