+++ packages/commonwealth/client/styles/pages/user_dashboard/user_dashboard_row.scss
@@ -1,19 +1,15 @@
 @import 'shared';
 
 .UserDashboardRow {
+  border-bottom-width: 2px;
   display: flex;
   flex-direction: column;
-  padding: 17px 22px;
-  margin-right: 60px;
-  border-bottom-width: 2px;
-
-  @include mediumSmallInclusive {
-    margin-right: 0;
-  }
+  gap: 16px;
+  padding: 16px 24px;
 
   &:hover {
     background: $neutral-50;
-    border-radius: 15px;
+    border-radius: $border-radius-corners;
   }
 
   .activity-content {
@@ -78,78 +74,6 @@
         -webkit-line-clamp: 2;
         -webkit-box-orient: vertical;
         margin: 7px 0 8px 0px;
-      }
-    }
-  }
-
-  .subscribe-button {
-    background: $neutral-50;
-    border: 1px solid $purple-400 !important;
-    color: $purple-400;
-    cursor: pointer;
-
-    svg {
-      stroke: $purple-400;
-    }
-  }
-
-  .icon-row {
-    margin-top: 10px;
-    margin-bottom: 10px;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-
-    &-events {
-      margin-left: 0px;
-    }
-
-    &-left {
-      display: flex;
-      margin-left: 20px;
-
-      @include mediumSmallInclusive {
-        display: none;
-      }
-
-      // button {
-      //   border: 1px solid $neutral-100;
-      //   margin-right: 25px;
-
-      //   @include smallInclusive {
-      //     // For temporary break point behavior
-      //     margin-top: 5px;
-      //   }
-
-      //   &:hover {
-      //     background: $neutral-50;
-      //     border: 1px solid $purple-400;
-      //     color: $purple-400;
-      //     cursor: pointer;
-
-      //     svg {
-      //       stroke: $purple-400;
-      //     }
-      //   }
-      // }
-    }
-
-    &-right {
-      svg {
-        stroke: $neutral-600;
-      }
-
-      button {
-        border: none;
-        color: $neutral-600;
-        background: transparent;
-      }
-    }
-
-    .cui-menu {
-      button {
-        border: none;
-        margin: 0;
       }
     }
   }
